+++ src/routes.ts
@@ -5,6 +5,7 @@ export default [
     route("search", "pages/SearchPage.tsx"),
     route("chat/:id?", "pages/ChatPage.tsx"),
     route("dispatcher/run", "pages/DispatcherRunPage.tsx"),
+    route("dataplayer_legacy_run", "pages/DataplayerLegacyRunPage.tsx"),
     route("dataplayer", "pages/DataplayerPage.tsx"),
     route("*", "pages/NotFound.tsx"),
 ] satisfies RouteConfig;
